ai/requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
torch==2.3.0
torchvision==0.18.0
grad-cam==1.5.0
